cdk/pyproject.toml
[project]
name = "kernelworx-cdk"
version = "0.1.0"
description = "AWS CDK infrastructure for Popcorn Sales Manager"
requires-python = ">=3.14"
dependencies = [
    "aws-cdk-lib>=2.215.0",
    "aws-cdk.aws-lambda-python-alpha>=2.215.0a0",
    "boto3>=1.41.5",
    "constructs>=10.0.0,<11.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "boto3-stubs[essential]>=1.41.5",
    "isort>=7.0.0",
    "moto>=5.1.17",
    "mypy>=1.19.0",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]

[tool.ruff]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
]
markers = [
    "integration: integration tests",
    "unit: unit tests",
]

[tool.coverage.run]
omit = [
    # cdk_stack.py requires AWS credentials for Route53 lookups during synthesis
    # It cannot be unit tested; it requires integration tests with real AWS
    "cdk/cdk_stack.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
